Start loading images sooner
